requirements.txt
oracledb
rich
python-dotenv
faker
pandas
